Move OutputHighlighter logic to more generic Highlighter and update the logic to preserve passed syntax highlighting from Prism
