+++ components/atoms/Highlighter/HighlighterStyles.ts
@@ -2,10 +2,6 @@ import styled from 'styled-components'
 
 export const StyledCodeBlock = styled.div`
     background-color: #e7e7e7;
-
-    span {
-        color: #000;
-    }
 `
 export const StyledCodeWrapper = styled.pre`
     background-color: #e7e7e7 !important;
@@ -15,4 +11,7 @@ export const StyledCodeWrapper = styled.pre`
 export const StyledHighlighterMatch = styled.mark`
     background-color: #ffdb45 !important;
     color: #000 !important;
+`
+export const StyledNonHighlighterMatch = styled.span`
+    color: #000;
 `
