+++ posts/getting-started-with-type-hinting-in-python.md
@@ -61,7 +61,7 @@ input='mypy sum.py'
 
 Running Mypy on the code will return the following errors.
 
-<OutputHighlighter
+<Highlighter
 input='sum.py:4: error: Argument 1 to "sum_two_numbers" has incompatible type "str"; expected "int"
 sum.py:4: error: Argument 2 to "sum_two_numbers" has incompatible type "str"; expected "int"
 Found 2 errors in 1 file (checked 1 source file)'
@@ -69,7 +69,7 @@ Found 2 errors in 1 file (checked 1 source file)'
 
 Code written with type hinting issues will not return any errors. If we put the first version of the `sum_two_numbers()` function that used integer types, we would instead receive the following output from Mypy.
 
-<OutputHighlighter
+<Highlighter
 input='Success: no issues found in 1 source file'
 />
