+++ posts/how-to-install-sourcegraph-on-linux-with-docker.md
@@ -111,7 +111,7 @@ This command is telling your computer to install and run Sourcegraph in port 708
 
 Once you run the command, you’ll receive some output indicating that Sourcegraph is getting set up. Sourcegraph will be ready when you receive the output similar to the following:
 
-<OutputHighlighter
+<Highlighter
 input='✱ Sourcegraph is ready at: http://127.0.0.1:7080'
 />
 
@@ -135,7 +135,7 @@ This command is telling your server to install and run Sourcegraph in port 7080 
 
 Once you run the command, you’ll receive some output indicating that Sourcegraph is getting set up. Sourcegraph will be ready when you receive output similar to the following. Your server’s IP address will display in place of `your-server-ip-address`.
 
-<OutputHighlighter
+<Highlighter
 input='✱ Sourcegraph is ready at: http://your-server-ip-address:7080'
 matcher='your-server-ip-address'
 />
