+++ posts/how-to-install-sourcegraph-on-mac-with-docker.md
@@ -62,7 +62,7 @@ This command is telling your computer to install and run Sourcegraph in port 708
 
 Once you run the command, you’ll receive some output indicating that Sourcegraph is getting set up. Sourcegraph will be ready when you receive the output similar to the following:
 
-<OutputHighlighter
+<Highlighter
 input='✱ Sourcegraph is ready at: http://127.0.0.1:7080'
 />
 
@@ -87,7 +87,7 @@ This command is telling your computer to install and run Sourcegraph in port 708
 
 Once you run the command, you’ll receive some output indicating that Sourcegraph is getting set up. Sourcegraph will be ready when you receive output similar to the following:
 
-<OutputHighlighter
+<Highlighter
 input='✱ Sourcegraph is ready at: http://127.0.0.1:7080'
 />
