+++ posts/how-to-stage-and-unstage-files-in-git.md
@@ -247,7 +247,7 @@ language='bash'
 
 After running the above command, all changes will be unstaged but will remain in the working directory. The output gives you a summary of those changes, as in the example below.
 
-<OutputHighlighter
+<Highlighter
 input='Unstaged changes after reset:
 M       index.html
 D       script.js
