+++ posts/how-to-troubleshoot-linux-permission-denied.md
@@ -13,7 +13,7 @@ type: posts
 
 If you are working on a Linux terminal and you receive the following output, you are probably trying to work with a file where the permissions have not been properly configured.
 
-<OutputHighlighter
+<Highlighter
 input='<filename>: Permission denied'
 />
 
@@ -48,7 +48,7 @@ language='bash'
 
 Depending on what else is in your working directory, you may have several lines of output, but you should be able to identify the file we just created, and review output for that line that is similar to the following.
 
-<OutputHighlighter
+<Highlighter
 input='-rw-r--r--  1 your-user your-user    12 Oct 21 23:10 test-file.txt'
 matcher='your-user'
 />
@@ -90,7 +90,7 @@ language='bash'
 
 Now, review the updated file permissions, using the `ls -l` command as before. You will receive output next to that file that is similar to the following. 
 
-<OutputHighlighter
+<Highlighter
 input='--w-------  1 your-user your-user    12 Oct 21 23:10 test-file.txt'
 matcher='your-user'
 />
@@ -108,7 +108,7 @@ language='bash'
 
 Once you enter the above command, you'll receive the `Permission denied` error message.
 
-<OutputHighlighter
+<Highlighter
 input='cat: test-file.txt: Permission denied'
 />
 
@@ -129,7 +129,7 @@ matcher='test-file'
 
 To verify that the file has now has read permissions, we'll use `ls -l`. This should display output similar to the following.
 
-<OutputHighlighter
+<Highlighter
 input='-rw-r--r--  1 your-user your-user    12 Oct 21 23:10 test-file.txt'
 matcher='your-user'
 />
@@ -146,7 +146,7 @@ language='bash'
 
 The command should now display the contents of the file to your terminal.
 
-<OutputHighlighter
+<Highlighter
 input='Hello World'
 />
 
@@ -170,7 +170,7 @@ Like with many situations in Linux, using the magic word `sudo` will allow you t
 
 We'll be continuing with the original example, `test-file.txt`, with the following permissions.
 
-<OutputHighlighter
+<Highlighter
 input='--w-------  1 your-user your-user    12 Oct 21 23:10 test-file.txt'
 matcher='your-user'
 />
@@ -183,7 +183,7 @@ matcher='test-file'
 language='bash'
 />
 
-<OutputHighlighter
+<Highlighter
 input='cat: test-file.txt: Permission denied'
 matcher='test-file'
 />
@@ -200,7 +200,7 @@ language='bash'
 
 Using the `sudo` command will open the file up to us. 
 
-<OutputHighlighter
+<Highlighter
 input='Hello World'
 />
